refactor(js): extract browser Flare assembly into side-effect-free src/browser.ts
